Apply improvements from code review
